requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Werkzeug==3.0.2
gunicorn==20.1.0
jinja2==3.1.4